+++ .gitignore
@@ -1,3 +1,4 @@
+*.iml
 .env
 /jsonnet/vendor
 ./jsonnet/vendor
