--- jsonnet/environments/default/spec.json
@@ -1,15 +1,0 @@
-{
-  "apiVersion": "tanka.dev/v1alpha1",
-  "kind": "Environment",
-  "metadata": {
-    "name": "environments/prod",
-    "namespace": "grafana-operator-system"
-  },
-  "spec": {
-    "apiServer": "https://api.play.gepaplexx.com:6443",
-    "namespace": "grafana-operator-system",
-    "resourceDefaults": {},
-    "expectVersions": {},
-    "injectLabels": true
-  }
-}
