--- jsonnet/generate.sh
@@ -1,3 +1,0 @@
-rm -r ./manifests
-tk export manifests environments/default/main.jsonnet
-rm ./manifests/manifest.json
