--- jsonnet/lib/k.libsonnet
@@ -1,1 +1,0 @@
-import 'github.com/jsonnet-libs/k8s-libsonnet/1.24/main.libsonnet'
