--- jsonnet/manifests/apps-v1.Deployment-ns-proxy.yaml
@@ -1,45 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  labels:
-    tanka.dev/environment: ef8004c80176691d8a0f3688c077aa401e0b6083174fa72e
-  name: ns-proxy
-  namespace: grafana-operator-system
-spec:
-  minReadySeconds: 10
-  replicas: 1
-  revisionHistoryLimit: 10
-  selector:
-    matchLabels:
-      name: ns-proxy
-  template:
-    metadata:
-      labels:
-        name: ns-proxy
-    spec:
-      containers:
-      - env:
-        - name: ADMIN_GROUP
-          value: Gepaplexx
-        - name: CLIENT_SECRET
-          value: TO-BE-SET
-        - name: DEV
-          value: "false"
-        - name: KEYCLOAK_CERT_URL
-          value: https://sso.apps.play.gepaplexx.com/realms/internal/protocol/openid-connect/certs
-        - name: LOG_LEVEL
-          value: info
-        - name: TENANT_LABEL
-          value: namespace
-        - name: TOKEN_EXCHANGE
-          value: "false"
-        - name: UPSTREAM_BYPASS_URL
-          value: https://thanos-querier.openshift-monitoring.svc.cluster.local:9091
-        - name: UPSTREAM_URL
-          value: http://prom-label-proxy.grafana-operator-system.svc.cluster.local:9095
-        image: ghcr.io/lucostus/namespace-proxy:sha-2d87a7b
-        imagePullPolicy: IfNotPresent
-        name: ns-proxy
-        ports:
-        - containerPort: 8080
-          name: http
