--- jsonnet/manifests/apps-v1.Deployment-prom-label-proxy.yaml
@@ -1,40 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  labels:
-    tanka.dev/environment: ef8004c80176691d8a0f3688c077aa401e0b6083174fa72e
-  name: prom-label-proxy
-  namespace: grafana-operator-system
-spec:
-  minReadySeconds: 10
-  replicas: 1
-  revisionHistoryLimit: 10
-  selector:
-    matchLabels:
-      name: prom-label-proxy
-  template:
-    metadata:
-      labels:
-        name: prom-label-proxy
-    spec:
-      containers:
-      - args:
-        - --insecure-listen-address=0.0.0.0:9095
-        - --upstream=https://thanos-querier.openshift-monitoring.svc.cluster.local:9091
-        - --label=namespace
-        - --query-param=namespace
-        - --enable-label-apis
-        - --error-on-replace
-        image: ghcr.io/lucostus/prom-label-proxy:multi-value-regex-v0.0.4
-        imagePullPolicy: IfNotPresent
-        name: prom-label-proxy
-        ports:
-        - containerPort: 9095
-          name: http
-        volumeMounts:
-        - mountPath: /etc/ssl/certs/
-          name: openshift-service-ca
-      volumes:
-      - configMap:
-          name: openshift-service-ca.crt
-        name: openshift-service-ca
