--- jsonnet/manifests/v1.Service-ns-proxy.yaml
@@ -1,16 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  labels:
-    name: ns-proxy
-    tanka.dev/environment: ef8004c80176691d8a0f3688c077aa401e0b6083174fa72e
-  name: ns-proxy
-  namespace: grafana-operator-system
-spec:
-  ports:
-  - name: ns-proxy-http
-    port: 8080
-    targetPort: 8080
-  selector:
-    name: ns-proxy
-  type: ClusterIP
