--- jsonnet/manifests/v1.Service-prom-label-proxy.yaml
@@ -1,16 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  labels:
-    name: prom-label-proxy
-    tanka.dev/environment: ef8004c80176691d8a0f3688c077aa401e0b6083174fa72e
-  name: prom-label-proxy
-  namespace: grafana-operator-system
-spec:
-  ports:
-  - name: prom-label-proxy-http
-    port: 9095
-    targetPort: 9095
-  selector:
-    name: prom-label-proxy
-  type: ClusterIP
