--- jsonnet/manifests/v1.ServiceAccount-ns-proxy.yaml
@@ -1,8 +1,0 @@
-apiVersion: v1
-automountServiceAccountToken: true
-kind: ServiceAccount
-metadata:
-  labels:
-    tanka.dev/environment: ef8004c80176691d8a0f3688c077aa401e0b6083174fa72e
-  name: ns-proxy
-  namespace: grafana-operator-system
